Decompose page into components
